requirements.txt
captcha==0.2.1
Django==1.8
django-appconf==1.0.2
django-bootstrap3==8.1.0
django-easy-pdf==0.1.0
django-extensions==1.7.5
django-fsm==2.4.0
django-fsm-admin==1.2.4
django-fsm-log==1.3.0
django-graphos==0.3.23
olefile==0.44
Pillow==4.0.0
six==1.10.0
